requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
diskcache==5.6.3
docker-pycreds==0.4.0
emoji==2.14.1
frozenlist==1.5.0
gitdb==4.0.12
gitpython==3.1.44
gql==3.5.2
graphql-core==3.2.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
mcp==1.6.0
mdurl==0.1.2
multidict==6.3.2
numpy==2.2.4
packaging==24.2
platformdirs==4.3.7
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pydantic==2.11.2
pydantic-core==2.33.1
pydantic-settings==2.8.1
pygments==2.19.1
python-dotenv==1.1.0
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
sentry-sdk==2.25.1
setproctitle==1.3.5
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.1
tenacity==9.1.2
typer==0.15.2
typing-extensions==4.13.1
typing-inspection==0.4.0
urllib3==2.3.0
uuid-utils==0.10.0
uvicorn==0.34.0
wandb==0.19.9
weave==0.51.41
yarl==1.19.0
